requirements.txt
streamlit
beautifulsoup4
pdfminer.six
sentence-transformers
faiss-cpu
langchain>=0.2.0
huggingface_hub
llama-cpp-python>=0.2.90
tqdm